agents/transport/requirements.txt
strands-agents>=0.1.0
bedrock-agentcore>=0.1.0
boto3>=1.34.0
requests>=2.31.0
aws-opentelemetry-distro>=0.10.1
